More test reorganization  (and fixture simplification).
