+++ tests/impure/README.md
@@ -1,0 +1,5 @@
+# Impure tests
+
+These tests require isolation of global state, and cannot be run in parallel.
+
+* Tests which require a repository are in the [_with_repo/_](with_repo/README.md) subdirectory.
