+++ tests/impure/test_xgit.py
@@ -10,8 +10,7 @@ def test_xsh(with_xgit):
     '''
     Test the xsh proxy.
     '''
-    def _t(*_, XSH: XonshSession, **__):
-        assert isinstance(XSH.env, MutableMapping), \
-            f"XSH.env not a MutableMapping {XSH.env!r}"
-    with_xgit(_t)
+    XSH = with_xgit.XSH
+    assert isinstance(XSH.env, MutableMapping), \
+        f"XSH.env not a MutableMapping {XSH.env!r}"
