+++ tests/impure/with_repo/README.md
@@ -1,0 +1,12 @@
+# Tests requiring a repository
+
+These are further divided into tests requiring a writable vs read-only repository.
+
+* [writeable/](writeable/README.md)
+* [read_only/](read_only/README.md)
+
+These tests use a zipped up repository, unzipped to a temporary directory. The scope is `"package"` for [read_only/](read_only/README.md) and `"function"` for _writeable/_.
+
+Data for the tests lives in:
+
+* [data/](data/README.md)
