+++ tests/impure/with_repo/data/.gitignore
