+++ tests/impure/with_repo/data/README.md
@@ -1,0 +1,9 @@
+# Data for repository tests
+
+This directory holds data used in the tests.
+
+Currently, there is a single data file.
+
+* **test_repo.zip**: Simple zipped repo, quick to unzip.
+
+The repository is zipped and unzipped via the module [zip_repo.py](../zip_repo.py).
