+++ tests/impure/with_repo/read_only/README.md
@@ -1,0 +1,3 @@
+# Read-only repository tests
+
+The tests in this directory share a single read-only repository. This is currently the same repository data used for the [writeable/](../writeable/README.md) tests, but as it is not repeatedly unpacked, a larger, more complex repository can reasonably be used if required to support test cases.
