+++ tests/impure/with_repo/writeable/README.md
@@ -1,0 +1,7 @@
+# Writeable repository tests
+
+The tests in this directory each have their own copy of the repository data.
+
+As the repository data is unzipped for each test, the repository used should be kept small.
+
+Tests which do not need read/write access should go in [read_only/](../read_only/README.md)
