+++ tests/pure/README.md
@@ -1,0 +1,3 @@
+# Pure unit tests
+
+These tests require no special cleanup and can be executed in parallel.
