+++ tests/pure/test_table.py
